misc changes
-fixed a text_widget string formatting issue
-changed the controller combo for advanced settings to back+d_up
-fixed an issue with cartographer profiles not being signed out.

+++ xlive/Blam/Engine/interface/new_hud_draw.cpp
@@ -561,7 +561,7 @@ void draw_hud_text_get_string(s_draw_hud_widget_input_results* widget_function_r
 
 	if(text_widget->flags.test(text_widget_flag_string_is_a_number))
 	{
-		real32 value = draw_hud_widget_get_value(-1, text_widget->string);
+		int32 value = (int32)draw_hud_widget_get_value(-1, text_widget->string);
 		if (text_widget->flags.test(text_widget_flag_force_2digit_number))
 		{
 			if (value < 0)
